Split infer.ts file into infer-subst and infer-solver.ts
